requirements.txt
appdirs==1.4.4
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.4.0
black==21.7b0
cfgv==3.3.1
click==8.0.3
colorama==0.4.4
distlib==0.3.4
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.7.0
djangorestframework==3.11.1
filelock==3.4.2
gunicorn==20.0.4
identify==2.4.2
iniconfig==1.1.1
isort==5.5.2
more-itertools==8.12.0
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
pluggy==0.13.1
pre-commit==2.15.0
psycopg2==2.8.6
py==1.11.0
pyparsing==3.0.6
pytest==6.0.1
pytest-django==3.9.0
python-decouple==3.4
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
virtualenv==20.13.0
